Fix a bug in attribute matching for generic protocols.
Matching a class (rather than an instance of a class) against a generic
protocol was broken. When the generic protocol is defined in the same file as
the class being matched against it, this is fixed by updating
abstract_utils.get_type_parameter_substitutions so it can get type parameter
values from classes as well as instances. When the classes are in different
files, load_lazy_attribute() needs to take type parameter values into account
when converting the protocol to an abstract value.

PiperOrigin-RevId: 386914661