Add __dataclass_fields__ member to dataclass overlay and Dataclass protocol.
This can be used to write type annotations that match any dataclass/instance of a dataclass, but not other classes/instances of other classes. Also supports a generic argument so that, for example, a dataclass whose keys are all strs can be matched. See the docstring of pytype_extensions.Dataclass for more details.

Also fix a bug in abstract_utils.get_signatures.

rechen: thank you very much for your help with this cl.
PiperOrigin-RevId: 387652898