Fix a bug in the pytd printer.
We sometimes spawn a new printer inside the printer, and the new printer didn't
always have the same settings as the old one. This led to a bug where, when
there existed a class called Union in a file, we sometimes mistakenly printed
Union instead of typing.Union when we meant the latter.

PiperOrigin-RevId: 387698183