Fix a few failing opensource tests.
The line number of an error on a decorated class with no parents changes
from one Python version to the next; sometimes the error is on the same line as
the decorator, and other times it's on the class 'def' line. Making the class
redundantly inherit from object gets around this.

PiperOrigin-RevId: 390204142