Merge tests/py3/ into tests/.
Now that we no longer have a tests/py2/ directory, we don't need tests/py3/
either. This change was generated by:
1. If there exists no tests/test_foo.py, then tests/py3/test_foo.py is moved to
   tests/test_foo.py.
2. If there does exist a tests/test_foo.py, then tests/test_foo.py is renamed
   to tests/test_foo1.py and tests/py3/test_foo.py is moved to
   tests/test_foo2.py.
I also made the corresponding BUILD and CMakeLists.txt changes.

I confirmed that the total number of targets and the number of test targets is
the same before and after this change.

PiperOrigin-RevId: 391410706