Fix typeshed _load_third_party_packages().
typeshed/tests/pytype_test was still broken for third-party stubs. This turned
out to be because, even ignoring @python2, a package can contain multiple
modules, and a module name can be provided by multiple packages. I also noticed
that typeshed seems to have added a @tests directory; accidentally including it
as a module name doesn't hurt but is technically wrong, so I filtered it out.

PiperOrigin-RevId: 392084509