Narrow the types of local variables based on test assertions.
Treats `unittest.TestCase` methods `assertIsNotNone` and `assertIsInstance` as
equivalent to type assertions, so that after e.g. `assertIsInstance(x, Foo)`, `x`
has its type set to `Foo`.

This narrowing is only done if `x` is a local variable in the current scope;
globals and closure cells will not be affected.

PiperOrigin-RevId: 394076667