Fix a bug in pytype's handling of late annotations.
We needed to wrap apply_annotation's call to eval_expr() in a
generate_late_annotations() context to avoid spurious name errors. Once we do
this, we do not need to explicitly stringify annotations.

'from __future__ import annotations' also does not become the default in 3.8.
(I'm pretty sure I was the one who incorrectly assumed this when originally
adding support for this import...oops.) In fact, it's now unclear if it'll ever
become the default, since it was taken out of 3.10.

PiperOrigin-RevId: 395281518