Slightly improve pytype's performance on a trivial file.
I noticed that pytype was taking a long time to analyze a very simple file
(about 2s for pytype/test_data/simple.py), so I profiled it to see if there
were any easy performance improvements we could make. I didn't find much, but
the changes here appear to speed pytype up by 5-10% on a trivial file, and
they're pretty harmless (we're already using Cython), so we might as well check
them in.

PiperOrigin-RevId: 395324530