Support arbitrary python values for typing.Annotated in pyi files.
Support added for python literals and function calls (for type constructors).
Function calls are converted to dicts, and then all annotations are converted
to their repr strings to be stored in pytd.Annotated nodes.

Annotations are not yet preserved for round-tripping.

Also moves around some code to fix dependency issues.

PiperOrigin-RevId: 396696017