Fix a timeout.
For some reason, setting the pyval of an existing starstarargs dict to {}
causes pytype to timeout on a fairly short file. Creating a new abstract dict
and copying over the type parameters seems to fix things.

I added a shortened version of the problematic file to test_data/perf/ and also
took the chance to delete some old test files that aren't referenced anywhere.

PiperOrigin-RevId: 398286654