Update pytype and typing annotation stubs for features added in attrs 21.1.0
These features are:

* attr's [Next Generation APIs](https://www.attrs.org/en/stable/api.html#next-gen) (`@define`, `@field`, etc)
* Generate an `__attrs_init__` if `init=False` was given to the class decorator.

Note for type annotations added in this CL:

Now that pytype is Python 3.6 and up only, it seems safe to use type annotations now.
I didn't go full on type annotations, just the ones in or near code I was editing, and that didn't require changing any dependencies among files.

PiperOrigin-RevId: 401056906