Move the vm.join_* methods into the context.
These are typegraph-related methods, so move them to where we've put the rest
of the typegraph stuff.

I didn't move the convert.merge_* methods and vm.make_native_function after all
because, on second thought, the context doesn't really seem like the right
place for helper methods for abstract values.

PiperOrigin-RevId: 401362930