Get rid of unnecesssary uses of BaseValue.isinstance_* methods.
The only modules in which these methods need to be used are abstract_utils,
class_mixin, mixin, and function, since the abstract module depends on them.

I also changed the PropertyInstance, ClassMethod, and StaticMethod special
builtins to inherit from abstract.Function so that code that does isinstance
checks on Function doesn't need to separately check for these builtins.

PiperOrigin-RevId: 401382243