Add test base support for multi-file tests.
Allows setting up a dependency tree before checking a file. Python files in the
dependency tree are analysed in order, and the corresponding pyi files
generated. The pythonpath for all files is set to the root of the dependency tree.

Also converts one test file as a more extended test of the new functionality.

PiperOrigin-RevId: 406256121