Add a feature flag, --gen-stub-imports, to improve pyi import handling.
Currently, pytype adds "x: module" constants to pyi files, and all attributes
on builtins.module are typed as Any, so accessing whatever_module.x.attribute
always produces Any. This is fixed by instead adding "import x" statements and
changing the pyi parser to always generate aliases for import statements.

This is a very large change because of all the changes to test output and the
places we need to plumb through --gen-stub-imports so that the pyi parser sees
it. The actually interesting changes are in:
* pytype/output.py
* pytype/pyi/modules.py
* pytype/pytd/printer.py
* pytype/pytd/pytd_visitors.py

PiperOrigin-RevId: 406940492