Whittle down vm.py further so we can enable pytype.
* Moves some more methods that were in vm.py only because we wanted them in a
  central location into more sensible locations (mostly context.py).
* Creates a new vm_utils.py module and moves a bunch of vm.py's private methods
  into it. I basically just read through vm.py and moved anything that looked
  like it didn't do a large amount of reading/writing VM state.

With these changes, the amount of time it takes pytype to analyze vm.py is
reduced to ~10s for inference and 15-20s for checking, so we can remove the
`# skip-file` directive. The vm.py/vm_utils.py split is sort of haphazard,
which might make navigation a little harder, but having cross-references should
make navigation considerably easier, so I think we still come out ahead.

PiperOrigin-RevId: 407699280