Allow disables for invalid calls in 3.9 to be put on the same lines as in 3.7.
This makes it easier to upgrade from Python 3.7 to 3.9.

This change also fixes a slightly annoying issue where we didn't allow function
call disables to go after the closing parenthesis if there was nothing else on
that line, so you had to write

def f(
  x  # pytype: disable=wrong-arg-types
)

rather than

def f(
  x
)  # pytype: disable=wrong-arg-types

There's one downside to this change, which is that in 3.8+, it applies a
disable to a function even if it's not on the final line, so

def f(
  x,
  y,  # pytype: disable=wrong-arg-types
  z,
)

disables wrong-arg-types for f. This also means that disable behavior in 3.7
and 3.8+ is slightly different. This may be fixable with some extra parsing
and/or version-specific logic in directors.py, but that's more troublesome to
maintain, so I'm inclined to just let this difference exist unless it starts
causing bugs and/or problems with the upgrade.

PiperOrigin-RevId: 410392112