Remove pytype's ability to accept python_version=3.
This has been useless since we dropped target Python 2 support. Instead,
python_version should be omitted or passed in as None.

There are a few places in this change where I had to write
'python_version or sys.version_info[:2]'. I didn't pull this out into a helper
method because all but one of these will disappear in follow-up changes that
will pass in config.Options rather than python_version in a lot of places.

PiperOrigin-RevId: 411119902