Split out abstract.BaseValue into its own file.
This moves BaseValue into a new file, _base.py, along with
abstract_utils.get_template and compute_template, which are only used in
BaseValue. This has several advantages:
* We avoid putting anything except abstract classes in abstract.py because it
  is already so bloated; if we split it up, we can put helper methods and
  classes close to where they're used.
* xrefs and type-checking will actually work in _base.py.
* Module boundaries may help with reinforcing good conceptual boundaries (e.g.,
  BaseValue shouldn't really reference its subclasses).

All the code in _base.py is moved from other files, but I don't know a way to
display the diff properly, so I left a few comments with screenshots instead.

PiperOrigin-RevId: 411878308