Put 'duplicate top-level identifier' fix behind a feature flag.
This seems to have caused quite a few breakages. As far as I can tell, it's
because we now handle the 'bidict.bidict' class properly, which is a good
thing, but this has also exposed some sort of weird bug involving overloads and
__init__. I'd like to get a PyPI release out today, so I'm going to stick this
behind a flag for now and dig into it more next week.

For https://github.com/google/pytype/issues/1057.

PiperOrigin-RevId: 412129190