Remove pytd optimizations that broke overloads.
A couple of the optimizations in pytype.pytd.optimize were messing up function
overloads in .py files. These optimizations aren't very helpful anymore, as
far as I can tell (they were probably more useful back when pytype regularly
inferred multiple signatures for a single function), so the easiest thing to do
is to get rid of them.

One of the removed optimizations was hiding a bug in the RemoveDuplicates
visitor by removing signatures that the latter failed to detect as duplicate,
so I fixed that bug as well.

PiperOrigin-RevId: 414802324