--allow-recursive-types: Add intermediate variables for parameterized types.
Previously, when a recursive type was parameterized like so:
```
X = Union[T, List['X']]
Y = X[int]
```
the parameterized type was being written out as-is to the type stub (so
"Y = X[int]" would appear literally in the stub). This caused problems with
stub validation (because pytype doesn't expect types like pytd.Union to appear
as the base type of a pytd.GenericType) and parsing (the parser would try to
evaluate "X[int]"), in addition to the output being difficult to get correct
(things get weird when a type is parameterized multiple times).

The simplest way I could think of to fix these issues is to instead create
intermediate variables to hold parameterized types, like so:
```
X = Union[T, List[X]]
Y = _X_LBAR_int_RBAR
_X_LBAR_int_RBAR = Union[int, List[_X_LBAR_int_RBAR]]
```

This requires (1) tweaking sub_one_annotation to construct an accurate
expression string for LateAnnotations it creates, (2) figuring out which
LateAnnotations are parameterized recursive types and writing those out as
module-level variables, and (3) coming up with a way to transform
LateAnnotation expressions into variable names and back.

This is the last piece of the initial implementation of recursive types! :D

PiperOrigin-RevId: 421477310