Support `typing.Final` without an explicit type.
```
x: Final = 10
```
means "infer the value of `x`, and tag it `Final`".

Refactors a bit to move `Final` from the annotation type to metadata within
`annotation_utils` rather than returning it as an extracted type from
`apply_annotation`. This introduces a bit of tech debt because we should
ideally be doing this in other code paths too (see `vm.byte_STORE_SUBSCR()`).

PiperOrigin-RevId: 422969779