Replace periods with _DOT_ in recursive type names.
When a recursive type is forwarded, it can end up with a dotted name that needs
to be flattened to something like Mod1_DOT_Type. If it is forwarded again, its
name can end up as Mod2.Mod1_DOT_Type, so unflattening has to handle both
regular and flattened periods.

PiperOrigin-RevId: 425707502