Avoid triggering PYTHONBYTESWARNING=2 errors.
Setting PYTHONBYTESWARNING=2 causes `str(some_bytes_object)` to fail. Pytype
has never intentionally avoided calling str() on a bytes object, but we happen
to not do it most of the time, so tests that set PYTHONBYTESWARNING=2 and run
pytype get lucky and pass. The latest typeshed update causes them to fail by
introducing a bytes Literal into the stdlib stubs. We don't own the tests in
question, so the easiest way to fix this is to instead use
`repr(some_bytes_object)`, which produces the same output.

PiperOrigin-RevId: 427908070