Fix a bug in generate_opcode_diffs and add a missing 3.10 opcode.
The script missed an opcode (RERAISE) whose index changed in 3.10.

Technically speaking, we should be using different class definitions for
RERAISE in pyc/opcodes.py depending on the target version, but using the 3.10
definition in older versions doesn't seem to cause any issues, so it's simpler
to just always use the newest definition.

For #1022.

PiperOrigin-RevId: 429463279