Add an overlay for clif.python.type_customization.extend.
I have an upcoming change to support nested classes in pytype, which as a side
effect fixes a bug where decorated classes wouldn't get analyzed if the
decorator isn't recognized as returning the class. This in turn generates
errors in a bunch of @extend-decorated classes, unless we add an overlay.

Note that this change does *not* add full support for
type_customization.extend; it only gets rid of the most obvious errors.

PiperOrigin-RevId: 432300381