Add a --strict-primitive-comparisons flag.
Adds a feature flag for the builtins.pytd change in
https://github.com/google/pytype/commit/b10ccb1727a46b3abf3fe4a36a47e7550d39b42d,
since it causes ~80 new breakages. (I didn't bother to put the slots.py change
behind the flag because it causes <20 breakages.)

Modifies the pyi parser to recognize `if PYTYPE_OPTIONS.flag_name` checks so
that feature flags can be used to modify pyi contents.

PiperOrigin-RevId: 432983156