Merge collections.namedtuple() and typing.NamedTuple() builders
Refactors both classes to expose an identical "construct a named tuple from
function arguments" workflow, and extracts argument handling code into a common
base class and some helper structs.

One side effect is that collections.namedtuples are "promoted" to
typing.NamedTuples, which have a slightly richer interface; if that turns out
to be a problem we can pass in an argument to distinguish between them but in
practice it shouldn't matter.

Also cleans up the comments and docstrings a bit.

PiperOrigin-RevId: 438381586