Do a better job of determining whether a typing(_extensions) import is allowed.
The immediate reason for this change is to fix a test that was failing in 3.10
because we were failing to log [not-supported-yet] for
typing_extensions.TypeGuard. I took the opportunity to simplify some things in
typing_overlay and typing_extensions_overlay to make future development easier
and enable us to log more accurate and helpful error messages.

For #1022.

PiperOrigin-RevId: 438436286