Simplify function arguments before SimpleFunction uses them.
There was a comment in SimpleFunction.call() explaining why the arguments were
deliberately not simplified, but simplifying them doesn't break any of our
tests (in fact, it catches a mistake in one of the tests!), so I just deleted
the comment.

This fixes a few more namedtuple-related errors at head.

PiperOrigin-RevId: 439432265