Fix a method binding bug.
The is_subclass check in _lookup_from_mro_flat needs to use the original lookup
class, not the base class that we're currently looking for the attribute on.
Since I don't want to pass the original class to _lookup_from_mro_flat, I moved
the part that needs access to the class back into lookup_from_mro.

PiperOrigin-RevId: 441893419