Add code structure for gradual migration to argument-by-argument call matching.
This adds some code that will let our current function call matching algorithm
and a new, faster one exist side-by-side until the new one can handle all the
cases that the current one does. It also enables the new matching for trivial
cases (when we're matching against an InterpreterFunction with no TypeVars in
the function signature, the containing class (for a method), or the function
arguments). This is already enough to show a pretty impressive speedup in our
function call benchmark but not in the real-world use case I tested. If
anything, it's a tad slower because of the _is_generic_call check, but that'll
go away eventually.

PiperOrigin-RevId: 446044514