Add a *skipped* test for [] (Generics are too lenient)
I'm really just interested in fixing this for protos, and in theory if we hard-code `message.Message` in the matcher then we might want to add a proto-specific regression test. But for now, this can be written with general Python classes.

Without the _skip_, the failure is:
```
AssertionError: Errors not found:
Line 13: wrong-arg-types
Line 27: wrong-arg-types
```
PiperOrigin-RevId: 455681387