Internal copy of GitHub PR #1237.
pyupgrade --py37-plus **/*.py

Python f-strings are shorter, easier to read, and faster. https://www.scivision.dev/python-f-string-speed

This PR touches both internal and external files, so I merged it directly, but
we also need to import it so that the internal changes aren't reverted the next
time we do an export.

PiperOrigin-RevId: 457566465