Adjust opcode line numbers for return statements in python 3.10+
Python 3.10 changed the way line numbers are assigned to `return` statements
within `with` blocks, to attach to the `with` statement. This follows the
actual control flow in the python interpreter, but makes it hard to report
errors attached to the right line in the source code.

We examine the AST and use the line number information from the parser to
adjust the `RETURN_VALUE` opcodes to have the same line numbers as the source
`return` statements that generated them. `RETURN_VALUE` opcodes corresponding
to implicit `return None` continue to be moved to the last line of the
function.

PiperOrigin-RevId: 460639960