Add more logging to tracer_vm.py and fix a couple of issues this reveals.
I was trying to track down a performance issue that occurred at some point
during the inference step, and it took me forever to pin down because of how
sparse the logging was. This change adds log statements to important substeps
of tracer_vm.compute_types(). It also removes a bit of dead code and fixes a
bug in pytd_classes_for_call_traces, both of which I noticed because they were
causing log statements to never be hit.

PiperOrigin-RevId: 465200574