Add a matcher.BadMatch data structure.
This allows us to store some information in the return value of bad_matches
that we previously had to pass around separately. I renamed the BadParam class
to BadType, since I reused it in BadMatch and it doesn't make much sense to
have a reference to a "Param" there. The only user-visible effect of this
change is slightly improved error messages, due to bad_matches now being able
to return a postprocessed version of the expected type.

PiperOrigin-RevId: 465211620