Change how we detect infinite recursion in __new__ calls.
My previous attempt caused __new__ calls to no longer be recorded in
InterpreterFunction._call_records, which then led to some weird inference
results. Instead, this change does more fine-grained tracking of what stage of
instance creation vm.init_class() is on (instantiating = calling __new__,
initializing = calling __init__) so that we can avoid calling __new__ again if
we see that we're already in the instantiating stage. Using classes as keys in
the cache exposed a bug in LiteralClass.__eq__, which I also fixed.
PiperOrigin-RevId: 476010028