Add `Attrs` protocol and `__attrs_attrs__` member to attr overlay.
This can be used to write type annotations that match any `Attrs` class or instance of an `Attrs` class, but not other classes/instances of other classes. Also supports a generic argument so that, for example, a dataclass whose keys are all strs can be matched. See the docstring of pytype_extensions.Attrs for more details.

PiperOrigin-RevId: 476486826