Make creating and copying AliasingDicts less error-prone.
I discovered that first setting the `uf` attribute and then setting keys on an
AliasingDict produces different behavior from first setting the keys and then
setting `uf`. This seems like an excellent way to have hard-to-find bugs, so
I made `uf` read-only, changed the `__init__` and `copy` methods to allow
easily specifying `uf` when creating a new AliasingDict, and renamed `uf` to
`aliases` for readability.

PiperOrigin-RevId: 478141861