Add an optional no-return-any flag to disallow Any as a return type.
Also adds a mechanism to allow per-file opting in to feature flags, via a
file-level `features` directive, e.g. in this case
```
# pytype: features=no-return-any
```
would set the same option as `--no-return-any` as a command line argument,
overriding the CLI flag if both are supplied. The `features` directive can only
be used to enable a feature flag.

PiperOrigin-RevId: 485392928