Detect TypeGuard misuse.
Detects various forms of TypeGuard misuse. This resolves all our existing TODOs
related to TypeGuard, but I realized that I've missed a couple cases that need
supporting (TypeGuard in pyi files and in Callable), so we can't remove the
not-supported-yet error yet.

I deleted two of the TODOs without implementing them:
* I didn't switch init_annotation to init_class because I wasn't able to
  construct a test case in which this made a difference.
* I didn't optimize the code to avoid creating unnecessary new instances
  because this turned out to be more difficult/finicky than I expected. We can
  revisit this if we actually observe performance problems caused by TypeGuard.

PiperOrigin-RevId: 487726625