Suppress the convert cache when converting a paramspec return type.
Fixes a bug where multiple calls to the same decorator all get the first
substitution of type params.

This works but it introduces some technical debt; we should ideally figure out
why constant_to_value is not doing the right thing in this one case rather than
patching around it.

PiperOrigin-RevId: 492407544