Don't let one TypeVar contain another, and keep invalid TypeVars when possible.
I found user code that is doing stuff like:
  T1 = TypeVar('T1')
  T2 = TypeVar('T2', T1)
=/ Pytype was just silently downgrading types to Any. It's better to warn the
user that this is invalid. However, converting to Any as soon as we notice an
invalid TypeVar causes follow-on errors when attempting to use that TypeVar in
a generic class definition. As long as we have the TypeVar name, we can instead
create an unbounded, unconstrained TypeVar.

PiperOrigin-RevId: 492571068