Add a base class for a PyTDClass with one custom method.
builtins.type.__new__ and subprocess.Popen.__new__ have similar custom
implementations. This change creates a new base class for them both to inherit
from. For now, the _METHOD_NAME attribute is unnecessary (since it's always
__new__), but in the upcoming typeshed update, typeshed switches to defining
Popen.__init__, so we'll soon need this attribute. This will also make it
easier to add a custom implementation of abc.ABCMeta.__new__, which again will
be needed for updating typeshed.

PiperOrigin-RevId: 492589488