Don't report container errors for late types.
We can't report the "not a container" pyi error when a class has a LateType as
a base, since the LateType could resolve to a container type.

With this change, it's possible that we'll miss genuine container errors, but
we're up to four user reports of false positives (three bugs, plus a direct
message this morning).

PiperOrigin-RevId: 493425492