Add pytype:load_pytd to the :api target and reduce usages of :libvm.
Out of curiosity, I looked through all of the usages of pytype:libvm and
realized that, for the vast majority of them, the only module not in the public
API that was needed was load_pytd. I added :load_pytd to :api and removed all
the usages of :libvm that I could. (While there's nothing terribly wrong with
targets under pytype/... depending on :libvm, it's still nice to make the
dependency structure clearer and to see how close/far :api is to providing a
usable interface.)

PiperOrigin-RevId: 501430070