Fix type error in devtools/python/pytype:cfg_visualizer.
I also made the generate_visualization function in cfg_visualizer private. It
took me a few minutes to find the right replacement for the old generate
function, and I think it would have been more obvious if it was easier to tell
that generate_visualization was a private helper called by the two public
generate_* functions.

PiperOrigin-RevId: 522488788