Add new Variable method PasteBindingWithNewData.
Adds a method to add a binding to a variable with origins (but not data) copied
from another binding. The motivation for this is that when I tried to implement
literal erasure in containers, I discovered that changing PasteBinding(b) to
AddBinding(new_data, {b}, node) can cause a performance regression, presumably
due to adding another layer of origins. Directly copying the original binding's
origins fixes this.

I initially tried to make AddBinding smart enough to auto-magically detect when
it should copy origins, but this didn't work because it requires you to have a
reference to the node that the old binding was added at. I considered adding an
extra parameter to AddBinding or PasteBinding for the new behavior, but that
seemed more complicated and harder to understand than adding a new method.

I reviewed all the existing AddBinding calls in the pytype code and changed
them to use PasteBindingWithNewData where appropriate. While doing this, I also
noticed that some call sites should have been using PasteBinding/PasteVariable,
so I fixed those as well.

PiperOrigin-RevId: 529469217