Mark JUMP_BACKWARD and JUMP_BACKWARD_NO_INTERRUPT as NO_NEXT.
I'm a little confused by our NO_NEXT flag because I don't see it reflected in
CPython, but JUMP_FORWARD has it, and adding it to JUMP_BACKWARD fixes a crash.

After this change, when pytype -V3.11 is run on:
```
for _ in range(10):
  pass
```
we're left with a spurious `No attribute '__iter__' on int` error, which I
assume is because the CALL opcode used for range() isn't implemented yet.

PiperOrigin-RevId: 550056403