Split init_class() into two methods depending on whether we need the new node.
All init_class() usages except two only need the created instance. We were
ending up with a lot of `_, instance = init_class()` or code that forwarded the
node just because it was there. I created a new init_class_and_forward_node()
for the only two callers that need the node (based on functional tests and
presubmits) and changed init_class() to only return the instance.

PiperOrigin-RevId: 551633919