Handle unexpected property call failures more gracefully.
A call to a property should never fail. Even if we annotate `self` as something
weird, pytype tries to catch that early. But I managed to hit a weird corner
case that I can't reproduce locally, where a property call raises
WrongArgType(). This causes a frame to be pushed and never popped, which leads
to a crash when pytype attempts to pop the frame underneath. This took me a
long time to debug.

This change makes future debugging hopefully easier by:
1) Catching the failed function call when it happens.
2) Making sure a pushed frame is always popped no matter what.

PiperOrigin-RevId: 551951594