Fix how pytype handles aliases to properties in stubs.
Previously, when looking up aliases, pytype was grabbing the last signature it
saw. For properties with setters and/or deleters, this did the wrong thing, but
I guess no one ever noticed?

With the recent changes to add decorator support, pytype sees that the aliased
function has a decorator that references <unknown name>.setter and raises a bad
dependency error.

PiperOrigin-RevId: 555665439