Make Overlay._convert_member pass a module name to members.
Previously, _convert_member only passed the context when building a member, but
PyTDFunction.make expects a module name, so we always had to override or wrap
make() to pass one in. Additionally, changing the module name when reusing
overlays between modules was a lot of work. This new approach is overall
simpler, at the cost of sometimes having to `del module` in builders that don't
care about the module name.

I renamed `overlay.build` to the more descriptive `overlay.add_name`, so that I
could add another `overlay.drop_module` helper.

PiperOrigin-RevId: 557916078