pyi parser: split ast->pytd generation into discrete steps.
Splits pytd generation into three steps: (1) conversion of constants from
ast.Constant to types.Pyval in a new visitor that runs before
_GeneratePytdVisitor, (2) conversion of type annotations to pytd in
_GeneratePytdVisitor.enter(), and (3) conversion of the rest of the ast tree to
pytd in _GeneratePytdVisitor.visit(). I also made _GeneratePytdVisitor raise an
exception if it ever encounters a node that it doesn't know how to visit, to
ensure that the assumption that a node's children have already been visited
when the node itself is visited is not violated.

Ideally, we would also pull type annotation conversion into a separate visitor,
but it requires knowledge of type aliases, which _GeneratePytdVisitor builds up
incrementally as it visits nodes.

This makes the parser code easier to reason about because we know at any
particular step exactly what state the tree should be in.

This change also contains a few follow-on improvements that the split enables:
* Name aliasing is supported more consistently.
* _GeneratePytdVisitors converts decorators to aliases as soon as it enters
  function definitions, rather than partially converting them and finishing the
  conversion in post-processing.
* Unexpected literals are detected as soon as they are visited, rather than the
  detection being delayed until the construction of a parameterized type.
* visit_Assign reuses visit_AnnAssign and separates handling of pytd types and
  non-pytd types like types.SlotDecl more cleanly.

PiperOrigin-RevId: 558988833