Use pycnite for pyc file reading.
Gets rid of pyc.magic and pyc.loadmarshal and imports those functions from
pycnite instead.

Also fixes the pyc file writer in pyc.compile_bytecode to match the cpython
output (we were compensating for it in the reader earlier).

NOTE: This currently defaults to supporting 3.8-3.10 bytecode; pycnite has
support for 3.11 bytecode but we have not integrated it yet.
PiperOrigin-RevId: 564796132