Track decorator names.
This will come in useful for moving around errors produced when applying
decorators based on Python version. Even if that doesn't pan out, this is good
information to keep around and cleans up the code a bit by combining the
'decorators' and 'is_decorated' attributes. Unrecognized decorators are
filtered out in output.py to minimize behavioral changes.

I also increased the shard count for a test that (probably unrelatedly)
happened to time out once while I was testing this change.

PiperOrigin-RevId: 568621934