Do not treat await calls as loops in 3.11
In python 3.11, `await f()` compiles to

```
LOAD f
CALL
GET_AWAITABLE
SEND <post-await>
YIELD_VALUE
JUMP_BACKWARD_NO_INTERRUPT <send opcode>
<post-await>
```

but we do not want to treat it as a loop since that messes up data stack
unification at the `SEND` and `<post-await>` points. Instead we just use the
`SEND` call to mark `<post-await>` as reachable and elide the jumps.

(In 3.10 this compiles to straight-line code with `YIELD_FROM`)

PiperOrigin-RevId: 570813589