Don't propagate block locals when a block cannot be exited from.
For code like this:
```
if __random__:
  x = 3
else:
  raise SomeException()
print(x)
```
Python 3.10 generates separate bytecode sequences for taking the first branch
and printing x, then taking the second branch and printing x, while 3.11
combines the two. This exposes a bug in the block graph where it thinks that x
may be undefined when it's being printed because it's not defined in the `else`
branch. Fixed by marking the `else` block as a "dead end"; i.e., its outgoing
edges are ignored.
PiperOrigin-RevId: 570907836