Change how we detect `with` and `async for` blocks in _add_setup_except.
I disassembled a bunch of code snippets, and in all of them, the exception
table entries for `async for` targeted an END_ASYNC_FOR opcode, and the ones
for `with` had `lasti` set. Filtering the exception table according to these
guessed rules and processing only the first entry for each line seems to fix
the exception-related bug in `test_flow1` and to make some of the code for
handling jumps unnecessary. I built a couple of our broken
python_static_analysis.pynext targets, and sadly this doesn't fix them, but it
doesn't introduce any new crashes at least.

The test_flow1 test was failing in both 3.11 and 3.8. Since 3.8 is relatively
old now, I just skipped the test in 3.8.

PiperOrigin-RevId: 571165899