Small fix to `parse_pyi`.
`parse_pyi.py` currently raises an exception when trying to decompose
the results of `parser.parse_pyi`. It only returns a single value,
but the left-hand side of the assignment specifies two variables.

Since the second variable is unused, all this CL does is remove that
portion of the assignment statement.

#tftypes

PiperOrigin-RevId: 572626607