Re-implement the SEND opcode.
The way this is supposed to work in 3.11 is that SEND gets a value from the
generator, YIELD_VALUE yields it, and JUMP_BACKWARDS_NO_INTERRUPT jumps back to
the SEND, repeatedly, until there are no more values to yield. At that point,
SEND pushes the generator's return value onto the stack and jumps past the
backwards jump. However, we removed the jump opcode, so we can't follow this
implementation. Reusing 3.10's YIELD_FROM implementation seems to work.

I also renamed some of the variables in our generator-related VM code to make
it easier to follow.

PiperOrigin-RevId: 572628644