Remove extraneous check in RenameModuleVisitor.
`pytd_visitors.RenameModuleVisitor` previously skipped
renaming a module if, when matching against the text of
a node, it detected a `.` character in any text trailing
the previous module name.

This prevents the visitor from correctly removing module
names added during pytype's serialization of ASTs.

This change removes that check and only that check. The
remaining check (which verifies that there is no text
preceding the match) should be sufficient to prevent
erroneous replacement in partial-match situations.

#tftypes

PiperOrigin-RevId: 572677520